domains.zyx
Key Points 
 Tricki was a small dog and was excessively loved by his mistress, Mrs. Pumphery. 
 He was greedy & loved food, but no exercise. 
 As a result he became fat & Mrs. Pumphery believed that he became lazy because he suffered from 
malnutrition. 
 Mrs Pumphery was a wealthy woman. She loved her dog too much. 
 She gave him some little extras between meals to build him up, malt, cod-liver oil and a bowl of Horlicks at 
night. 
 She gave him lot to eat but no exercise. 
 Once he became ill & she was worried & made a frantic call to Harriot who was a vet. 
 James already had given suggestions regarding Tricki’s balanced diet. But she never followed it. 
 He was a vet. He was shocked to see Tricki hugely fat. 
 He advised her to cut down on the sweet things. 
 He expected call came within a few days & it happened. He was called by Mrs. Pumphery as Tricki refused to 
eat. 
 He told her that Tricki needed immediate hospitalization for keeping under observation at his surgery for a 
fortnight. 
 There he gave him no food no medicine but plenty of water. 
 Gradually he mixed with other dogs & started playing with them. 
 He had a lot of exercises & fun at the surgery. 
 He was completely recovered & transformed into a hard muscled dog. 
 James Harriot treated the dog without medicine or surgery. 
 She was happy & filled with gratitude. 
 She thanked Harriot & said it was a triumph of surgery. 
Summary 
Mrs. Pumphrey was a rich and emotional lady who has a cute pet called Tricky. She loved him so much that she 
overfed him. Tricki had put on a lot of weight which made him lethargic. When doctor Herriot saw the fat dog like 
a bloated sausage he was shocked. He made a plan and told Mrs. Pumphrey that Tricki needed treatment for a 
ailment which was possible only in hospital. The doctor knew that Mrs. Pumphrey would not be able to keep the 
dog on a proper diet. When the greedy dog went to the hospital he had to remain hungry if he was slow to approach 
for food. He was given food at fixed intervals. Soon he shed off a lot of his body weight and became quite active. 
When Tricki was at home he was pampered by Mrs. Pumphrey. He had a luxurious life when he was served with 
cream cake chocolate and horlicks. But when he came to the doctor he became all right. When Mrs. Pumphrey saw 
her active dog she thanked the doctor and felt that it was a triumph of surgery. 
Bloated - swollen with fluid or gas; ailment - illness;